Add files via upload
The above project is the practical implementation of multinomial naive bayes theorem. It recommends most appropriate teaching technique to be used to teach an particular topic to a certain age group. It takes the topic ,the targeted age group(young, medium, old) and their approximate intellectual ability(high, medium, low) as input and suggest the most appropriate teaching technique